! STR$ was rounding its argument to the declared type before handing it over, which is a real loss and not a formality: PB evaluates a SINGLE-typed expression on the x87 at eighty bits and rounds when it is STORED, and the NAME - rt_str_f32 against rt_str_f64 - picks the significant-digit count, not the width of the arithmetic. The two share a body. Rounding here threw the bits away before the formatter saw them, and the PRINT path already knew better; STR$ now does the same thing for the same stated reason * invisible under plain pb35, where a SINGLE prints 7 digits either way and STR$(2 / 3) is ".6666667" before and after. Under $COMPAT tb10 the formatter prints seventeen and it showed: .6666666865348816 - the float 2/3 widened back up - against the direct emitter's .6666666666666667. Now .6666666666666666, one ulp out rather than eight digits out, and the remainder is a different cause: the IR folds float arithmetic at DOUBLE while the x87 computes at eighty bits, and IrConstantFloat has nowhere to put the extra bits * rt_str_from_single/double in the C runtime take a long double now, matching rt_print_single/double, which were corrected for this same reason earlier

+++ runtime/pbc_rt.h
@@ -68,8 +68,8 @@ void *rt_str_from_u16(uint16_t v);
 void *rt_str_from_i32(int32_t v);
 void *rt_str_from_u32(uint32_t v);
 void *rt_str_from_i64(int64_t v);
-void *rt_str_from_single(float v);
-void *rt_str_from_double(double v);
+void *rt_str_from_single(long double v);
+void *rt_str_from_double(long double v);
 void *rt_str_from_ext(long double v);
 
 /* MKx$ / CVx binary record encoders */
